Crate brontes_inspect

The brontes_inspect crate is designed to efficiently detect and analyze MEV. Emphasizing modularity and ease of use, this crate provides a robust foundation for developing custom inspectors, streamlining the process of MEV strategy identification.

By abstracting complex tasks such as decoding, normalization, metadata fetching, and price tracking, brontes_inspect allows developers to concentrate on the unique logic of their MEV detection strategies. This design philosophy ensures that users can easily integrate their own inspectors, tailored to specific MEV strategies, without delving into the underlying infrastructure details.

Inspector

Inspector is a trait defining a method process_tree. This method takes a BlockTree and Metadata as input and returns a vector of tuples, each containing a BundleHeader and a BundleData.

#[async_trait::async_trait]
pub trait Inspector: Send + Sync {
    async fn process_tree(
        &self,
        tree: Arc<BlockTree<Actions>>,
        metadata: Arc<Metadata>,
    ) -> Vec<Bundle>;
}

The BlockTree represents a block of classified & normalized Ethereum transactions & their traces, and the Metadata contains price information & relevant off chain data such as mempool data & centralized exchange price data relevant to that block. The process_tree method analyzes the block and identifies instances of the MEV strategy that the inspector is designed to detect.

Individual Inspectors

The brontes_inspect crate provides several individual inspectors, each designed to detect a specific type of MEV strategy. These inspectors are defined in their respective modules:

atomic_backrun

cex_dex

jit

sandwich

liquidations

long_tail

Each inspector implements the Inspector trait and provides its own implementation of the process_tree method.

Composer

The Composer is a special type of inspector that combines the results of individual inspectors to identify more complex MEV strategies. It takes an array of individual inspectors and a BlockTree and Metadata as input, running each inspector on the block and collecting their results.

pub struct Composer<'a, const N: usize> {
    inspectors_execution: InspectorFut<'a>,
    pre_processing:       BlockPreprocessing,
}

The Composer uses to define a filter that orders results from individual inspectors. This ensures that lower-level actions are composed before higher-level actions, which could affect the composition.

Additionally, the Composer provides a Future implementation for use in asynchronous contexts. When polled, it runs the individual inspectors in parallel and collects their results, processing them to identify complex MEV strategies.

In summary, the brontes_inspect crate offers tools for detecting and analyzing MEV strategies in Ethereum transactions. Individual inspectors identify specific MEV strategies, while the Composer combines these results to identify more complex strategies.

Modules

atomic_arb

cex_dex

This module implements the CexDexInspector, a specialized inspector designed to detect arbitrage opportunities between centralized exchanges (CEXs) and decentralized exchanges (DEXs).

composer

The composer module in brontes-inspect specializes in analyzing and processing MEV data. Its primary functions include composing complex MEV types from simpler ones and deduplicating overlapping MEV occurrences.

discovery

The DiscoveryInspector module in brontes-inspect specializes in identifying potential MEV transactions. It does this by looking for transactions that are x standard deviations above the average priority fee (where x is the std_dev_threshold paramater, set to 2 by default), or have a coinbase transfer, or are private transactions based on the indexed mempool transactions we have in our metadata database (s/o chainbound).

jit

liquidations

long_tail

sandwich

shared_utils

Macros

define_mev_precedence

Defines precedence rules among different MEV types for the purpose of deduplication.

mev_composability

Defines rules for composing multiple child MEV types into a single, complex parent MEV type.

Structs

InspectorsIter

An iterator over the variants of Inspectors

Enums

Inspectors

Traits

Inspector